# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Barber Cukur-in",
        "title": "Barber Cukur-in",
        "contact": {
            "name": "Nuryanto",
            "url": "https://www.linkedin.com/in/nuryanto-1b2721156/",
            "email": "nuryantofattih@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/barber-service/barber/barber": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barber"
                ],
                "summary": "GetList Barber",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PerPage",
                        "name": "perpage",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "InitSearch",
                        "name": "initsearch",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SortField",
                        "name": "sortfield",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseModelList"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barber"
                ],
                "summary": "Add Barber",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BarbersPost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
        "/barber-service/barber/barber/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barber"
                ],
                "summary": "GetById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barber"
                ],
                "summary": "Rubah Barber",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BarbersPost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barber"
                ],
                "summary": "Delete Barber",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
        "/barber-service/barber/beranda": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barber Beranda"
                ],
                "summary": "GetList Barber Beranda",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PerPage",
                        "name": "perpage",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "InitSearch",
                        "name": "initsearch",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SortField",
                        "name": "sortfield",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ParamView",
                        "name": "paramview",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseModelList"
                        }
                    }
                }
            }
        },
...
        "/barber-service/barber/notification/beranda": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "summary": "Jumlah Notif yg belum dibuka",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BarberCapsterPost": {
            "type": "object",
            "properties": {
                "capster_id": {
                    "type": "integer"
                }
            }
        },
        "models.BarberPaketPost": {
            "type": "object",
            "properties": {
                "paket_id": {
                    "type": "integer"
                }
            }
        },
        "models.BarbersPost": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "barber_capster": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BarberCapsterPost"
                    }
                },
                "barber_name": {
                    "type": "string"
                },
                "barber_paket": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BarberPaketPost"
                    }
                },
                "file_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "operation_end": {
                    "type": "string"
                },
                "operation_start": {
                    "type": "string"
                }
            }
        },
...
        "models.ResponseModelList": {
            "type": "object",
            "properties": {
                "all_column": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "define_column": {
                    "type": "string"
                },
                "define_size": {
                    "type": "string"
                },
                "last_page": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "tool.ResponseModel": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "description": "Code int         `json:\"code\"`",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}